Add files via upload
In that project, I exercised my understanding of how a hashtable functions to store key-value pairs. I implemented a hashtable with different probing methodologies to resolve collisions. I used my hashtable implementation as a storage for login credentials, username, and password, for a website. This was a service that all websites used to create and manage user accounts. The password was stored in encrypted form for security, which was an industry standard practice.